Factoring out is_... primitives
