+++ phylanx/include/ast.hpp
@@ -7,6 +7,9 @@
 #define PHYLANX_AST_HPP
 
 #include <phylanx/config.hpp>
+#include <phylanx/ast/detail/is_identifier.hpp>
+#include <phylanx/ast/detail/is_placeholder.hpp>
+#include <phylanx/ast/detail/is_placeholder_ellipses.hpp>
 #include <phylanx/ast/generate_ast.hpp>
 #include <phylanx/ast/match_ast.hpp>
 #include <phylanx/ast/node.hpp>
