+++ src/ast/node.cpp
@@ -94,86 +94,6 @@ namespace phylanx { namespace ast
     }
 
     ///////////////////////////////////////////////////////////////////////////
-    namespace detail
-    {
-        struct is_placeholder
-        {
-            template <typename Ast>
-            bool operator()(Ast const& ast) const
-            {
-                return ast::is_placeholder(ast);
-            }
-        };
-
-        struct is_placeholder_ellipses
-        {
-            template <typename Ast>
-            bool operator()(Ast const& ast) const
-            {
-                return ast::is_placeholder_ellipses(ast);
-            }
-        };
-
-        struct is_identifier
-        {
-            template <typename Ast>
-            bool operator()(Ast const& ast) const
-            {
-                return ast::is_identifier(ast);
-            }
-        };
-
-        struct identifier_name
-        {
-            template <typename Ast>
-            std::string operator()(Ast const& ast) const
-            {
-                return ast::identifier_name(ast);
-            }
-        };
-    }
-
-    bool is_placeholder(primary_expr const& pe)
-    {
-        return visit(detail::is_placeholder(), pe);
-    }
-
-    bool is_placeholder_ellipses(primary_expr const& pe)
-    {
-        return visit(detail::is_placeholder_ellipses(), pe);
-    }
-
-    bool is_identifier(primary_expr const& pe)
-    {
-        return visit(detail::is_identifier(), pe);
-    }
-
-    std::string identifier_name(primary_expr const& pe)
-    {
-        return visit(detail::identifier_name(), pe);
-    }
-
-    bool is_placeholder(operand const& op)
-    {
-        return visit(detail::is_placeholder(), op);
-    }
-
-    bool is_placeholder_ellipses(operand const& op)
-    {
-        return visit(detail::is_placeholder_ellipses(), op);
-    }
-
-    bool is_identifier(operand const& op)
-    {
-        return visit(detail::is_identifier(), op);
-    }
-
-    std::string identifier_name(operand const& op)
-    {
-        return visit(detail::identifier_name(), op);
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
     void serialize(hpx::serialization::input_archive& ar, optoken& id, unsigned)
     {
         int val;
